PLAT-355: rename token service -> authentication service
